#6 Rename files to kebab case
