+++ lib/services/transaction-manager.js
